+++ tasks.py
@@ -7,7 +7,7 @@ from compas_invocations2 import docs
 from compas_invocations2 import style
 from compas_invocations2 import tests
 from compas_invocations2 import grasshopper
-from invoke import Collection
+from invoke.collection import Collection
 
 ns = Collection(
     docs.help,