+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Updated minimum library version to `2.14.1` in Rhino8 GH components.
 * Changed name of YAK package from `bluejay` to `compas`.
 * Fixed broken `scaled()` method in `Sphere`, `Cylinder`, and `Capsule` classes by overriding to accept uniform scaling factor.
+* Fixed bug in `compas.geometry.PlanarSurface`.
 
 ### Removed